requirements.txt
pandas
openai
python-dotenv
notebook
ipykernel
streamlit
matplotlib